refactor: fix linting rules and rename some param name

+++ actors/actor_system.go
@@ -107,7 +107,7 @@ type ActorSystem interface {
 	// InCluster states whether the actor system has started within a cluster of nodes
 	InCluster() bool
 	// GetPartition returns the partition where a given actor is located
-	GetPartition(actorName string) uint64
+	GetPartition(actorName string) int
 	// Subscribe creates an event subscriber to consume events from the actor system.
 	// Remember to use the Unsubscribe method to avoid resource leakage.
 	Subscribe() (eventstream.Subscriber, error)
@@ -445,12 +445,12 @@ func (x *actorSystem) Unsubscribe(subscriber eventstream.Subscriber) error {
 }
 
 // GetPartition returns the partition where a given actor is located
-func (x *actorSystem) GetPartition(actorName string) uint64 {
+func (x *actorSystem) GetPartition(actorName string) int {
 	if !x.InCluster() {
 		// TODO: maybe add a partitioner function
 		return 0
 	}
-	return uint64(x.cluster.GetPartition(actorName))
+	return x.cluster.GetPartition(actorName)
 }
 
 // InCluster states whether the actor system is started within a cluster of nodes
